+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Product Backlog
-* [High] Remove Sequelize dependency from server/auth/local/passport.js and google/passport.js
+
+* [High] Change user.controller.js:upsert to use createUser()
+* [High] Setup nightly run of zero_old_passes() in PostgreSQL
 * [High] Remove connect-session-sequelize dependency from /server/config/express.js (might not need since no Twitter Oauth 1.0)
+* [High] Remove sequelize dependencies in integration tests
 * [High] Create DDL and remove seed.js from project
 * [Low] Remove Buffer.from() in user.controller.js:encryptPassword and reset passwords for local accounts
 * [Low] server/config/express.js - replace require() with imports
 * [Low] Add toast after forgotpassword
+
+<a name="1.0.15"></a>
+* [High] Remove Sequelize dependency from server/auth/local/passport.js and google/passport.js
 
 <a name="1.0.14"></a>
 * [Low] Added SendInBlue tags to order emails